# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CTU Theseus API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/classes": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "List Classes",
        "operationId": "list_classes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Create Class",
        "operationId": "create_class",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/classes/{class_id}": {
      "patch": {
        "tags": [
          "classes"
        ],
        "summary": "Update Class",
        "operationId": "update_class",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Class Id"
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "classes"
        ],
        "summary": "Delete Class",
        "operationId": "delete_class",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Class Id"
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/exports": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Create Export",
        "operationId": "create_export",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExportBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "export"
        ],
        "summary": "List Exports",
        "operationId": "list_exports",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Get Export",
        "operationId": "get_export",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassListResponse": {
        "properties": {
          "classes": {
            "items": {
              "$ref": "#/components/schemas/LabelClassOut"
            },
            "type": "array",
            "title": "Classes"
          }
        },
        "type": "object",
        "required": [
          "classes"
        ],
        "title": "ClassListResponse"
      },
      "ClassResponse": {
        "properties": {
          "class": {
            "$ref": "#/components/schemas/LabelClassOut"
          }
        },
        "type": "object",
        "required": [
          "class"
        ],
        "title": "ClassResponse"
      },
...
      "CreateClassBody": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "uiColorHex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uicolorhex"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateClassBody"
      },
      "CreateExportBody": {
        "properties": {
          "tier": {
            "type": "string",
            "enum": [
              "model",
              "devkit",
              "app"
            ],
            "title": "Tier"
          },
          "format": {
            "type": "string",
            "enum": [
              "onnx",
              "torchscript"
            ],
            "title": "Format"
          },
          "lang": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "python",
                  "typescript",
                  "csharp",
                  "java",
                  "pwa",
                  "flutter"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          }
        },
        "type": "object",
        "required": [
          "tier",
          "format"
        ],
        "title": "CreateExportBody"
      },
...
      "ExportAccepted": {
        "properties": {
          "exportId": {
            "type": "string",
            "format": "uuid",
            "title": "Exportid"
          }
        },
        "type": "object",
        "required": [
          "exportId"
        ],
        "title": "ExportAccepted"
      },
      "ExportListResponse": {
        "properties": {
          "exports": {
            "items": {
              "$ref": "#/components/schemas/ExportRow"
            },
            "type": "array",
            "title": "Exports"
          }
        },
        "type": "object",
        "required": [
          "exports"
        ],
        "title": "ExportListResponse"
      },
      "ExportResponse": {
        "properties": {
          "export": {
            "$ref": "#/components/schemas/ExportRow"
          }
        },
        "type": "object",
        "required": [
          "export"
        ],
        "title": "ExportResponse"
      },
      "ExportRow": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "runId": {
            "type": "string",
            "format": "uuid",
            "title": "Runid"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "tier": {
            "type": "string",
            "enum": [
              "model",
              "devkit",
              "app"
            ],
            "title": "Tier"
          },
          "format": {
            "type": "string",
            "enum": [
              "onnx",
              "torchscript"
            ],
            "title": "Format"
          },
          "lang": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "python",
                  "typescript",
                  "csharp",
                  "java",
                  "pwa",
                  "flutter"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "converting",
              "assembling",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "bundleKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundlekey"
          },
          "byteSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytesize"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "failedMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedmessage"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "readyAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readyat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "runId",
          "userId",
          "tier",
          "format",
          "lang",
          "status",
          "bundleKey",
          "byteSize",
          "checksum",
          "failedMessage",
          "createdAt",
          "readyAt",
          "updatedAt"
        ],
        "title": "ExportRow"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelClassOut": {
        "properties": {
          "classId": {
            "type": "string",
            "format": "uuid",
            "title": "Classid"
          },
          "datasetId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "uiColorHex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uicolorhex"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "classId",
          "datasetId",
          "name",
          "description",
          "uiColorHex",
          "isActive",
          "createdAt"
        ],
        "title": "LabelClassOut"
      },
...
      "UpdateClassBody": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "uiColorHex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uicolorhex"
          }
        },
        "type": "object",
        "title": "UpdateClassBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}